requirements.txt
Django==1.9.4
argparse==1.2.1
django-appconf==1.0.1
django-compressor==1.4
django-cors-headers==1.1.0
django-filter==0.11.0
djangorestframework==3.3.2
drf-nested-routers==0.11.1
mysqlclient==1.3.6
psycopg2==2.6.1
six==1.10.0
wsgiref==0.1.2
